Split EventBus.
