+++ src/core/CMakeLists.txt
@@ -3,20 +3,23 @@ set(NAME "Core")
 include(${PROJECT_SOURCE_DIR}/cmake/StaticAnalyzers.cmake)
 
 add_library(${NAME} STATIC
-  "Core/Log.cpp"
-  "Core/Log.hpp"
+  "Core/AppEvents.h"
   "Core/Application.cpp"
   "Core/Application.hpp"
-  "Core/Window.cpp"
-  "Core/Window.hpp"
-  "Core/DPIHandler.hpp"
-  "Core/Renderer.cpp"
-  "Core/ImGuiPass.cpp"
   "Core/AppUi.cpp"
+  "Core/DPIHandler.hpp"
+  "Core/EventBus.h"
+  "Core/EventBusImpl.h"
+  "Core/ImGuiPass.cpp"
   "Core/InputSystem.cpp"
-  "Core/SdlPathService.cpp"
+  "Core/Log.cpp"
+  "Core/Log.hpp"
+  "Core/Renderer.cpp"
   "Core/SDLContext.cpp"
   "Core/SDLContext.h"
+  "Core/SdlPathService.cpp"
+  "Core/Window.cpp"
+  "Core/Window.hpp"
 )
 
 # Define set of OS specific files to include
